Examples-Forms/requirements.txt
abstra
beautifulsoup4
boto3
colorlog
docxtpl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google.auth
invoice
jsonschema
lxml
numpy
openai
pandas
plotly
pydash
python-dateutil
requests
selenium
sendgrid-python
sheet2dict
simple-salesforce
slack-sdk
stripe
tabulate
unidecode
xlrd
black==23.1.0